+++ src/features/club-register/model/reducer/clubFormReducer.ts
@@ -1,18 +1,7 @@
 import validateField from '../../util/validateField';
-import { ClubFormData } from '../type';
+import { Action, StateProp } from '../type';
 
-type Action =
-  | { type: 'UPDATE_FIELD'; name: keyof ClubFormData; value: string }
-  | { type: 'VALIDATE_FIELD'; name: keyof ClubFormData }
-  | { type: 'UPDATE_LOGO'; file: File }
-  | { type: 'RESET_FORM' };
-
-type State = {
-  formData: ClubFormData;
-  errors: Partial<Record<keyof ClubFormData, string>>;
-};
-
-export const initialState: State = {
+export const initialState: StateProp = {
   formData: {
     name: '',
     category: '',
@@ -25,7 +14,7 @@ export const initialState: State = {
   errors: {},
 };
 
-export default function reducer(state: State, action: Action): State {
+export default function reducer(state: StateProp, action: Action): StateProp {
   switch (action.type) {
     case 'UPDATE_FIELD': {
       const { name, value } = action;
